+++ .github/CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ## 2020.16.0
 
+1. [A32NX/TCAS] Improve PFD Resolution Advisory graphics, introduce TCAS bus - @lukecologne (luke)
+
 ## 2020.15.0
 
 1. [OVHD] Implement FAULT light flickering during self-test sequence of PRIM, SEC, ELAC, FAC - @flogross89 (floridude)